+++ sdks/java/io/iceberg/src/main/java/org/apache/beam/sdk/io/iceberg/RecordWriterManager.java
@@ -25,7 +25,6 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
-import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 import org.apache.beam.sdk.schemas.Schema;
 import org.apache.beam.sdk.util.Preconditions;
@@ -91,6 +90,7 @@ class RecordWriterManager implements AutoCloseable {
     final Cache<PartitionKey, RecordWriter> writers;
     private final List<SerializableDataFile> dataFiles = Lists.newArrayList();
     @VisibleForTesting final Map<PartitionKey, Integer> writerCounts = Maps.newHashMap();
+    private final List<Exception> exceptions = Lists.newArrayList();
 
     DestinationState(IcebergDestination icebergDestination, Table table) {
       this.icebergDestination = icebergDestination;
@@ -113,11 +113,14 @@ class RecordWriterManager implements AutoCloseable {
                     try {
                       recordWriter.close();
                     } catch (IOException e) {
-                      throw new RuntimeException(
-                          String.format(
-                              "Encountered an error when closing data writer for table '%s', partition %s",
-                              icebergDestination.getTableIdentifier(), pk),
-                          e);
+                      RuntimeException rethrow =
+                          new RuntimeException(
+                              String.format(
+                                  "Encountered an error when closing data writer for table '%s', path: %s",
+                                  icebergDestination.getTableIdentifier(), recordWriter.path()),
+                              e);
+                      exceptions.add(rethrow);
+                      throw rethrow;
                     }
                     openWriters--;
                     dataFiles.add(SerializableDataFile.from(recordWriter.getDataFile(), pk));
@@ -195,7 +198,9 @@ class RecordWriterManager implements AutoCloseable {
 
   private final Map<WindowedValue<IcebergDestination>, List<SerializableDataFile>>
       totalSerializableDataFiles = Maps.newHashMap();
-  private static final Cache<TableIdentifier, Table> TABLE_CACHE =
+
+  @VisibleForTesting
+  static final Cache<TableIdentifier, Table> TABLE_CACHE =
       CacheBuilder.newBuilder().expireAfterAccess(10, TimeUnit.MINUTES).build();
 
   private boolean isClosed = false;
@@ -221,22 +226,28 @@ class RecordWriterManager implements AutoCloseable {
   private Table getOrCreateTable(TableIdentifier identifier, Schema dataSchema) {
     @Nullable Table table = TABLE_CACHE.getIfPresent(identifier);
     if (table == null) {
-      try {
-        table = catalog.loadTable(identifier);
-      } catch (NoSuchTableException e) {
+      synchronized (TABLE_CACHE) {
         try {
-          org.apache.iceberg.Schema tableSchema =
-              IcebergUtils.beamSchemaToIcebergSchema(dataSchema);
-          // TODO(ahmedabu98): support creating a table with a specified partition spec
-          table = catalog.createTable(identifier, tableSchema);
-          LOG.info("Created Iceberg table '{}' with schema: {}", identifier, tableSchema);
-        } catch (AlreadyExistsException alreadyExistsException) {
-          // handle race condition where workers are concurrently creating the same table.
-          // if running into already exists exception, we perform one last load
           table = catalog.loadTable(identifier);
+        } catch (NoSuchTableException e) {
+          try {
+            org.apache.iceberg.Schema tableSchema =
+                IcebergUtils.beamSchemaToIcebergSchema(dataSchema);
+            // TODO(ahmedabu98): support creating a table with a specified partition spec
+            table = catalog.createTable(identifier, tableSchema);
+            LOG.info("Created Iceberg table '{}' with schema: {}", identifier, tableSchema);
+          } catch (AlreadyExistsException alreadyExistsException) {
+            // handle race condition where workers are concurrently creating the same table.
+            // if running into already exists exception, we perform one last load
+            table = catalog.loadTable(identifier);
+          }
         }
+        TABLE_CACHE.put(identifier, table);
       }
-      TABLE_CACHE.put(identifier, table);
+    } else {
+      // If fetching from cache, refresh the table to avoid working with stale metadata
+      // (e.g. partition spec)
+      table.refresh();
     }
     return table;
   }
@@ -254,15 +265,7 @@ class RecordWriterManager implements AutoCloseable {
             icebergDestination,
             destination -> {
               TableIdentifier identifier = destination.getValue().getTableIdentifier();
-              Table table;
-              try {
-                table =
-                    TABLE_CACHE.get(
-                        identifier, () -> getOrCreateTable(identifier, row.getSchema()));
-              } catch (ExecutionException e) {
-                throw new RuntimeException(
-                    "Error while fetching or creating table: " + identifier, e);
-              }
+              Table table = getOrCreateTable(identifier, row.getSchema());
               return new DestinationState(destination.getValue(), table);
             });
 
@@ -283,6 +286,17 @@ class RecordWriterManager implements AutoCloseable {
       // removing writers from the state's cache will trigger the logic to collect each writer's
       // data file.
       state.writers.invalidateAll();
+      // first check for any exceptions swallowed by the cache
+      if (!state.exceptions.isEmpty()) {
+        IllegalStateException exception =
+            new IllegalStateException(
+                String.format("Encountered %s failed writer(s).", state.exceptions.size()));
+        for (Exception e : state.exceptions) {
+          exception.addSuppressed(e);
+        }
+        throw exception;
+      }
+
       if (state.dataFiles.isEmpty()) {
         continue;
       }