+++ sdks/python/apache_beam/examples/cookbook/bigtableio_it_test.py
@@ -25,7 +25,6 @@ import string
 import unittest
 import uuid
 from typing import TYPE_CHECKING
-from typing import List
 
 import pytest
 import pytz
@@ -53,7 +52,7 @@ except ImportError:
 if TYPE_CHECKING:
   import google.cloud.bigtable.instance
 
-EXISTING_INSTANCES: List['google.cloud.bigtable.instance.Instance'] = []
+EXISTING_INSTANCES: list['google.cloud.bigtable.instance.Instance'] = []
 LABEL_KEY = 'python-bigtable-beam'
 label_stamp = datetime.datetime.utcnow().replace(tzinfo=UTC)
 label_stamp_micros = _microseconds_from_datetime(label_stamp)